requirements.txt
watchdog==4.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
playwright>=1.42.0
schedule==1.2.0
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0